+++ petJournal/app/src/main/java/com/soujunior/petjournal/ui/ValidationEvent.kt
@@ -1,4 +1,4 @@
-package com.soujunior.petjournal.ui.accountManager.loginScreen
+package com.soujunior.petjournal.ui
 
 sealed class ValidationEvent {
     object Success : ValidationEvent()
